temporary fix for tests so that they pass

+++ tests/test_pipelines/test_rasterize.py
@@ -5,16 +5,12 @@ import pytest
 
 from eolearn.core import EONode, EOPatch, FeatureType, MapFeatureTask
 
-from eogrow.core.config import interpret_config_from_path
+from eogrow.core.config import collect_configs_from_path, interpret_config_from_dict
 from eogrow.pipelines.rasterize import RasterizePipeline
-from eogrow.utils.testing import ContentTester, check_pipeline_logs, create_folder_dict, run_and_test_pipeline
+from eogrow.utils.meta import load_pipeline_class
+from eogrow.utils.testing import check_pipeline_logs, compare_content
 
 pytestmark = pytest.mark.integration
-
-
-@pytest.fixture(scope="session", name="folders")
-def config_folder_fixture(config_folder, stats_folder):
-    return create_folder_dict(config_folder, stats_folder, "rasterize")
 
 
 class CropPreprocessTask(MapFeatureTask):
@@ -22,11 +18,8 @@ class CropPreprocessTask(MapFeatureTask):
 
     CROP_MAP = {"wheat": 1, "barley": 4, "cotton": 5, "sugar beet": 13, "grape": 14}
 
-    def map_method(self, dataframe, config):
-        crop_column = config.columns[0].values_column
-
-        dataframe[crop_column] = dataframe["CROP_TYPE"].apply(lambda crop_name: self.CROP_MAP[crop_name])
-
+    def map_method(self, dataframe):
+        dataframe["CROP_ID"] = dataframe["CROP_TYPE"].apply(lambda crop_name: self.CROP_MAP[crop_name])
         return dataframe
 
 
@@ -40,9 +33,7 @@ class CropRasterizePipeline(RasterizePipeline):
 
     def get_prerasterization_node(self, previous_node: EONode) -> EONode:
         """Applies mapping between names and crop ids"""
-        return EONode(
-            CropPreprocessTask(self.vector_feature, self.vector_feature, config=self.config), inputs=[previous_node]
-        )
+        return EONode(CropPreprocessTask(self.vector_feature, self.vector_feature), inputs=[previous_node])
 
 
 def add_vector_data(pipeline):
@@ -72,48 +63,67 @@ def add_vector_data(pipeline):
 
 
 @pytest.mark.parametrize("experiment_name", ["rasterize_pipeline_float"])
-def test_rasterize_pipeline(experiment_name, folders):
-    run_and_test_pipeline(experiment_name, **folders)
+def test_rasterize_pipeline(config_and_stats_paths, experiment_name):
+    config_path, stats_path = config_and_stats_paths("rasterize", experiment_name)
+    crude_configs = collect_configs_from_path(config_path)
+    raw_configs = [interpret_config_from_dict(config) for config in crude_configs]
+
+    first_run = True
+    for config in raw_configs:
+        pipeline = load_pipeline_class(config).from_raw_config(config)
+
+        if first_run:
+            filesystem = pipeline.storage.filesystem
+            folder = pipeline.storage.get_folder(pipeline.config.output_folder_key)
+            filesystem.removetree(folder)
+            first_run = False
+
+        pipeline.run()
+        check_pipeline_logs(pipeline)
+
+    compare_content(config_path, stats_path)
 
 
 @pytest.mark.chain
 @pytest.mark.order(before="test_rasterize_pipeline_features")
 @pytest.mark.parametrize("experiment_name", ["rasterize_pipeline"])
-def test_rasterize_pipeline_preprocess(folders, experiment_name):
-    # Can't use utility testing due to custom pipeline
-    config_filename = os.path.join(folders["config_folder"], experiment_name + ".json")
-    stat_path = os.path.join(folders["stats_folder"], experiment_name + ".json")
+def test_rasterize_pipeline_preprocess(config_and_stats_paths, experiment_name):
+    config_path, stats_path = config_and_stats_paths("rasterize", experiment_name)
+    crude_configs = collect_configs_from_path(config_path)
+    raw_configs = [interpret_config_from_dict(config) for config in crude_configs]
 
-    config = interpret_config_from_path(config_filename)
-    pipeline = CropRasterizePipeline.from_raw_config(config)
+    first_run = True
+    for config in raw_configs:
+        pipeline = CropRasterizePipeline.from_raw_config(config)
 
-    filesystem = pipeline.storage.filesystem
-    folder = pipeline.storage.get_folder(pipeline.config.output_folder_key)
-    filesystem.removetree(folder)
+        if first_run:
+            filesystem = pipeline.storage.filesystem
+            folder = pipeline.storage.get_folder(pipeline.config.output_folder_key)
+            filesystem.removetree(folder)
+            first_run = False
 
-    pipeline.run()
-    check_pipeline_logs(pipeline)
+        pipeline.run()
+        check_pipeline_logs(pipeline)
 
-    tester = ContentTester(filesystem, folder)
-    # tester.save(stat_path)
-    assert tester.compare(stat_path) == {}
+    compare_content(config_path, stats_path)
 
 
 @pytest.mark.chain
 @pytest.mark.parametrize("experiment_name", ["rasterize_pipeline_features"])
-def test_rasterize_pipeline_features(folders, experiment_name):
-    config_filename = os.path.join(folders["config_folder"], experiment_name + ".json")
-    stat_path = os.path.join(folders["stats_folder"], experiment_name + ".json")
+def test_rasterize_pipeline_features(config_and_stats_paths, experiment_name):
+    config_path, stats_path = config_and_stats_paths("rasterize", experiment_name)
+    crude_configs = collect_configs_from_path(config_path)
+    raw_configs = [interpret_config_from_dict(config) for config in crude_configs]
 
-    config = interpret_config_from_path(config_filename)
-    pipeline = RasterizePipeline.from_raw_config(config)
-    add_vector_data(pipeline)
+    first_run = True
+    for config in raw_configs:
+        pipeline = RasterizePipeline.from_raw_config(config)
 
-    pipeline.run()
-    check_pipeline_logs(pipeline)
+        if first_run:
+            add_vector_data(pipeline)
+            first_run = False
 
-    filesystem = pipeline.storage.filesystem
-    folder = pipeline.storage.get_folder(pipeline.config.output_folder_key)
-    tester = ContentTester(filesystem, folder)
-    # tester.save(stat_path)
-    assert tester.compare(stat_path) == {}
+        pipeline.run()
+        check_pipeline_logs(pipeline)
+
+    compare_content(config_path, stats_path)
